LuminAgents/requirements.txt
google-genai>=1.5.0
litellm>=1.35.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
requests>=2.31.0
python-dotenv>=1.0.0
python-telegram-bot>=21.0
streamlit>=1.32.0
aiosqlite>=0.20.0
pydantic>=2.6.0
aiohttp>=3.9.0
tavily-python>=0.3.0
